20 Oct, 2016 Test case for MapControl made.  In midst of debugging.

+++ StrangerThings/src/CITstrangerthings/control/MapControl.java
@@ -12,7 +12,7 @@ package CITstrangerthings.control;
 public class MapControl {
     
    
-    public int marbleCount(float userAnswer, float x, float y){
+    public int marbleCount(int userAnswer, int x, int y){
         
          if (userAnswer < 1 || userAnswer > 342) {
                     	return -1;
@@ -23,7 +23,7 @@ public class MapControl {
          if (y < 1 || y > 100) {
                     	return -1;
                              }
-         double compAns = ( ( (x/2) + y)/3)*2;
+         int compAns = (int)(( ( (x/2) + y)/3)*2);
  
          if (userAnswer != compAns) {
         	return 0;
